+++ tokenise.h
@@ -1,6 +1,6 @@
 #ifndef _TOKENISE_H_
 #define _TOKENISE_H_
 
-int tokenise_string(char* input_string, char delim, char ***tokens);
+size_t tokenise_string(char* input_string, char delim, char ***tokens);
 
 #endif
